@@ -77,8 +77,8 @@
     "@rollup/plugin-replace": "^2.3.3",
     "@testing-library/jest-dom": "^5.11.5",
     "@testing-library/jest-native": "^3.4.3",
-    "@testing-library/react": "^12.0.0",
-    "@testing-library/react-hooks": "^3.4.2",
+    "@testing-library/react": "https://pkg.csb.dev/testing-library/react-testing-library/commit/0e2cf7da/@testing-library/react/_pkg.tgz",
+    "@testing-library/react-hooks": "^7.0.1",
     "@testing-library/react-native": "^7.1.0",
     "@types/create-react-class": "^15.6.3",
     "@types/object-assign": "^4.0.30",
@@ -103,10 +103,10 @@
     "glob": "^7.1.6",
     "jest": "^26.6.1",
     "prettier": "^2.1.2",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
+    "react": "18.0.0-alpha-b9934d6db-20210805",
+    "react-dom": "18.0.0-alpha-b9934d6db-20210805",
     "react-native": "^0.64.1",
-    "react-test-renderer": "^16.14.0",
+    "react-test-renderer": "18.0.0-alpha-b9934d6db-20210805",
     "redux": "^4.0.5",
     "rimraf": "^3.0.2",
     "rollup": "^2.32.1",
