@@ -1,8 +1,8 @@
 /* eslint-disable valid-jsdoc, @typescript-eslint/no-unused-vars */
 import hoistStatics from 'hoist-non-react-statics'
-import React, { useContext, useMemo, useRef, useReducer } from 'react'
-import { isValidElementType, isContextConsumer } from 'react-is'
-import type { Store, Dispatch, Action, AnyAction } from 'redux'
+import React, { useCallback, useContext, useMemo } from 'react'
+import { isContextConsumer, isValidElementType } from 'react-is'
+import type { Dispatch, Store } from 'redux'
 
 import type {
   AdvancedComponentDecorator,
@@ -25,8 +25,6 @@ import defaultMapDispatchToPropsFactories from '../connect/mapDispatchToProps'
 import defaultMapStateToPropsFactories from '../connect/mapStateToProps'
 import defaultMergePropsFactories from '../connect/mergeProps'
 
-import { createSubscription, Subscription } from '../utils/Subscription'
-import { useIsomorphicLayoutEffect } from '../utils/useIsomorphicLayoutEffect'
 import shallowEqual from '../utils/shallowEqual'
 
 import {
@@ -34,6 +32,8 @@ import {
   ReactReduxContextValue,
   ReactReduxContextInstance,
 } from './Context'
+import { createReduxContext } from './Provider'
+import { useStoreSource } from '../utils/useStoreSource'
 
 // Define some constant arrays just to avoid re-creating these
 const EMPTY_ARRAY: [unknown, number] = [null, 0]
@@ -61,143 +61,6 @@ function storeStateUpdatesReducer(
   const [, updateCount] = state
   return [action.payload, updateCount + 1]
 }
-
-type EffectFunc = (...args: any[]) => void | ReturnType<React.EffectCallback>
-
-// This is "just" a `useLayoutEffect`, but with two modifications:
-// - we need to fall back to `useEffect` in SSR to avoid annoying warnings
-// - we extract this to a separate function to avoid closing over values
-//   and causing memory leaks
-function useIsomorphicLayoutEffectWithArgs(
-  effectFunc: EffectFunc,
-  effectArgs: any[],
-  dependencies?: React.DependencyList
-) {
-  useIsomorphicLayoutEffect(() => effectFunc(...effectArgs), dependencies)
-}
-
-// Effect callback, extracted: assign the latest props values to refs for later usage
-function captureWrapperProps(
-  lastWrapperProps: React.MutableRefObject<unknown>,
-  lastChildProps: React.MutableRefObject<unknown>,
-  renderIsScheduled: React.MutableRefObject<boolean>,
-  wrapperProps: unknown,
-  actualChildProps: unknown,
-  childPropsFromStoreUpdate: React.MutableRefObject<unknown>,
-  notifyNestedSubs: () => void
-) {
-  // We want to capture the wrapper props and child props we used for later comparisons
-  lastWrapperProps.current = wrapperProps
-  lastChildProps.current = actualChildProps
-  renderIsScheduled.current = false
-
-  // If the render was from a store update, clear out that reference and cascade the subscriber update
-  if (childPropsFromStoreUpdate.current) {
-    childPropsFromStoreUpdate.current = null
-    notifyNestedSubs()
-  }
-}
-
-// Effect callback, extracted: subscribe to the Redux store or nearest connected ancestor,
-// check for updates after dispatched actions, and trigger re-renders.
-function subscribeUpdates(
-  shouldHandleStateChanges: boolean,
-  store: Store,
-  subscription: Subscription,
-  childPropsSelector: (state: unknown, props: unknown) => unknown,
-  lastWrapperProps: React.MutableRefObject<unknown>,
-  lastChildProps: React.MutableRefObject<unknown>,
-  renderIsScheduled: React.MutableRefObject<boolean>,
-  childPropsFromStoreUpdate: React.MutableRefObject<unknown>,
-  notifyNestedSubs: () => void,
-  forceComponentUpdateDispatch: React.Dispatch<any>
-) {
-  // If we're not subscribed to the store, nothing to do here
-  if (!shouldHandleStateChanges) return
-
-  // Capture values for checking if and when this component unmounts
-  let didUnsubscribe = false
-  let lastThrownError: Error | null = null
-
-  // We'll run this callback every time a store subscription update propagates to this component
-  const checkForUpdates = () => {
-    if (didUnsubscribe) {
-      // Don't run stale listeners.
-      // Redux doesn't guarantee unsubscriptions happen until next dispatch.
-      return
-    }
-
-    const latestStoreState = store.getState()
-
-    let newChildProps, error
-    try {
-      // Actually run the selector with the most recent store state and wrapper props
-      // to determine what the child props should be
-      newChildProps = childPropsSelector(
-        latestStoreState,
-        lastWrapperProps.current
-      )
-    } catch (e) {
-      error = e
-      lastThrownError = e as Error | null
-    }
-
-    if (!error) {
-      lastThrownError = null
-    }
-
-    // If the child props haven't changed, nothing to do here - cascade the subscription update
-    if (newChildProps === lastChildProps.current) {
-      if (!renderIsScheduled.current) {
-        notifyNestedSubs()
-      }
-    } else {
-      // Save references to the new child props.  Note that we track the "child props from store update"
-      // as a ref instead of a useState/useReducer because we need a way to determine if that value has
-      // been processed.  If this went into useState/useReducer, we couldn't clear out the value without
-      // forcing another re-render, which we don't want.
-      lastChildProps.current = newChildProps
-      childPropsFromStoreUpdate.current = newChildProps
-      renderIsScheduled.current = true
-
-      // If the child props _did_ change (or we caught an error), this wrapper component needs to re-render
-      forceComponentUpdateDispatch({
-        type: 'STORE_UPDATED',
-        payload: {
-          error,
-        },
-      })
-    }
-  }
-
-  // Actually subscribe to the nearest connected ancestor (or store)
-  subscription.onStateChange = checkForUpdates
-  subscription.trySubscribe()
-
-  // Pull data from the store after first render in case the store has
-  // changed since we began.
-  checkForUpdates()
-
-  const unsubscribeWrapper = () => {
-    didUnsubscribe = true
-    subscription.tryUnsubscribe()
-    subscription.onStateChange = null
-
-    if (lastThrownError) {
-      // It's possible that we caught an error due to a bad mapState function, but the
-      // parent re-rendered without this component and we're about to unmount.
-      // This shouldn't happen as long as we do top-down subscriptions correctly, but
-      // if we ever do those wrong, this throw will surface the error in our tests.
-      // In that case, throw the error from here so it doesn't get lost.
-      throw lastThrownError
-    }
-  }
-
-  return unsubscribeWrapper
-}
-
-// Reducer initial state creation for our update reducer
-const initStateUpdates = () => EMPTY_ARRAY
 
 export interface ConnectProps {
   reactReduxForwardedRef?: React.ForwardedRef<unknown>
@@ -501,7 +364,7 @@ function connect<
   const wrapWithConnect: AdvancedComponentDecorator<
     TOwnProps,
     WrappedComponentProps
-  > = (WrappedComponent) => {
+  > = ((WrappedComponent) => {
     if (
       process.env.NODE_ENV !== 'production' &&
       !isValidElementType(WrappedComponent)
@@ -534,14 +397,42 @@ function connect<
       areMergedPropsEqual,
     }
 
-    // If we aren't running in "pure" mode, we don't want to memoize values.
-    // To avoid conditionally calling hooks, we fall back to a tiny wrapper
-    // that just executes the given callback immediately.
-    const usePureOnlyMemo = pure
-      ? useMemo
-      : (callback: () => void) => callback()
+    const useChildPropsSelector = (context: ReactReduxContextValue) => {
+      return useMemo(
+        () =>
+          defaultSelectorFactory(
+            context.store.dispatch,
+            selectorFactoryOptions
+          ),
+        [context.store]
+      )
+    }
 
-    function ConnectFunction<TOwnProps>(props: ConnectProps & TOwnProps) {
+    const useStateAndSubscribe = (
+      context: ReactReduxContextValue,
+      wrapperProps: Omit<WrappedComponentProps, 'reactReduxForwardedRef'>
+    ) => {
+      const childPropsSelector = useChildPropsSelector(context)
+      const getSnapshot = useCallback(
+        (store) => childPropsSelector(store.getState(), wrapperProps),
+        [childPropsSelector, wrapperProps]
+      )
+      return useStoreSource(context!.storeSource, getSnapshot)
+    }
+
+    const useDispatchOnly = (
+      context: ReactReduxContextValue,
+      wrapperProps: Omit<WrappedComponentProps, 'reactReduxForwardedRef'>
+    ) => {
+      const childPropsSelector = useChildPropsSelector(context)
+      return childPropsSelector(context.store.getState(), wrapperProps)
+    }
+
+    const useReduxContext = shouldHandleStateChanges
+      ? useStateAndSubscribe
+      : useDispatchOnly
+
+    function ConnectFunction(props: ConnectProps & TOwnProps) {
       const [propsContext, reactReduxForwardedRef, wrapperProps] =
         useMemo(() => {
           // Distinguish between actual "data" props that were passed to the wrapper component,
@@ -573,7 +464,9 @@ function connect<
         Boolean(props.store!.getState) &&
         Boolean(props.store!.dispatch)
       const didStoreComeFromContext =
-        Boolean(contextValue) && Boolean(contextValue!.store)
+        Boolean(contextValue) &&
+        Boolean(contextValue!.storeSource) &&
+        Boolean(contextValue!.store)
 
       if (
         process.env.NODE_ENV !== 'production' &&
@@ -588,159 +481,24 @@ function connect<
         )
       }
 
-      // Based on the previous check, one of these must be true
-      const store: Store = didStoreComeFromProps
-        ? props.store!
-        : contextValue!.store
+      const reduxContextValue: ReactReduxContextValue = useMemo(() => {
+        // Based on the previous check, one of these must be true
+        return didStoreComeFromProps
+          ? createReduxContext(props.store!)
+          : contextValue!
+      }, [didStoreComeFromProps, props.store, contextValue])
 
-      const childPropsSelector = useMemo(() => {
-        // The child props selector needs the store reference as an input.
-        // Re-create this selector whenever the store changes.
-        return defaultSelectorFactory(store.dispatch, selectorFactoryOptions)
-      }, [store])
-
-      const [subscription, notifyNestedSubs] = useMemo(() => {
-        if (!shouldHandleStateChanges) return NO_SUBSCRIPTION_ARRAY
-
-        // This Subscription's source should match where store came from: props vs. context. A component
-        // connected to the store via props shouldn't use subscription from context, or vice versa.
-        const subscription = createSubscription(
-          store,
-          didStoreComeFromProps ? undefined : contextValue!.subscription
-        )
-
-        // `notifyNestedSubs` is duplicated to handle the case where the component is unmounted in
-        // the middle of the notification loop, where `subscription` will then be null. This can
-        // probably be avoided if Subscription's listeners logic is changed to not call listeners
-        // that have been unsubscribed in the  middle of the notification loop.
-        const notifyNestedSubs =
-          subscription.notifyNestedSubs.bind(subscription)
-
-        return [subscription, notifyNestedSubs]
-      }, [store, didStoreComeFromProps, contextValue])
-
-      // Determine what {store, subscription} value should be put into nested context, if necessary,
-      // and memoize that value to avoid unnecessary context updates.
-      const overriddenContextValue = useMemo(() => {
-        if (didStoreComeFromProps) {
-          // This component is directly subscribed to a store from props.
-          // We don't want descendants reading from this store - pass down whatever
-          // the existing context value is from the nearest connected ancestor.
-          return contextValue!
-        }
-
-        // Otherwise, put this component's subscription instance into context, so that
-        // connected descendants won't update until after this component is done
-        return {
-          ...contextValue,
-          subscription,
-        } as ReactReduxContextValue
-      }, [didStoreComeFromProps, contextValue, subscription])
-
-      // We need to force this wrapper component to re-render whenever a Redux store update
-      // causes a change to the calculated child component props (or we caught an error in mapState)
-      const [[previousStateUpdateResult], forceComponentUpdateDispatch] =
-        useReducer(
-          storeStateUpdatesReducer,
-          // @ts-ignore
-          EMPTY_ARRAY as any,
-          initStateUpdates
-        )
-
-      // Propagate any mapState/mapDispatch errors upwards
-      if (previousStateUpdateResult && previousStateUpdateResult.error) {
-        throw previousStateUpdateResult.error
-      }
-
-      // Set up refs to coordinate values between the subscription effect and the render logic
-      const lastChildProps = useRef()
-      const lastWrapperProps = useRef(wrapperProps)
-      const childPropsFromStoreUpdate = useRef()
-      const renderIsScheduled = useRef(false)
-
-      const actualChildProps = usePureOnlyMemo(() => {
-        // Tricky logic here:
-        // - This render may have been triggered by a Redux store update that produced new child props
-        // - However, we may have gotten new wrapper props after that
-        // If we have new child props, and the same wrapper props, we know we should use the new child props as-is.
-        // But, if we have new wrapper props, those might change the child props, so we have to recalculate things.
-        // So, we'll use the child props from store update only if the wrapper props are the same as last time.
-        if (
-          childPropsFromStoreUpdate.current &&
-          wrapperProps === lastWrapperProps.current
-        ) {
-          return childPropsFromStoreUpdate.current
-        }
-
-        // TODO We're reading the store directly in render() here. Bad idea?
-        // This will likely cause Bad Things (TM) to happen in Concurrent Mode.
-        // Note that we do this because on renders _not_ caused by store updates, we need the latest store state
-        // to determine what the child props should be.
-        return childPropsSelector(store.getState(), wrapperProps)
-      }, [store, previousStateUpdateResult, wrapperProps])
-
-      // We need this to execute synchronously every time we re-render. However, React warns
-      // about useLayoutEffect in SSR, so we try to detect environment and fall back to
-      // just useEffect instead to avoid the warning, since neither will run anyway.
-      useIsomorphicLayoutEffectWithArgs(captureWrapperProps, [
-        lastWrapperProps,
-        lastChildProps,
-        renderIsScheduled,
-        wrapperProps,
-        actualChildProps,
-        childPropsFromStoreUpdate,
-        notifyNestedSubs,
-      ])
-
-      // Our re-subscribe logic only runs when the store/subscription setup changes
-      useIsomorphicLayoutEffectWithArgs(
-        subscribeUpdates,
-        [
-          shouldHandleStateChanges,
-          store,
-          subscription,
-          childPropsSelector,
-          lastWrapperProps,
-          lastChildProps,
-          renderIsScheduled,
-          childPropsFromStoreUpdate,
-          notifyNestedSubs,
-          forceComponentUpdateDispatch,
-        ],
-        [store, subscription, childPropsSelector]
-      )
+      const childProps = useReduxContext(reduxContextValue, wrapperProps)
 
       // Now that all that's done, we can finally try to actually render the child component.
       // We memoize the elements for the rendered child component as an optimization.
-      const renderedWrappedComponent = useMemo(
+      return useMemo(
         () => (
           // @ts-ignore
-          <WrappedComponent
-            {...actualChildProps}
-            ref={reactReduxForwardedRef}
-          />
+          <WrappedComponent {...childProps} ref={reactReduxForwardedRef} />
         ),
-        [reactReduxForwardedRef, WrappedComponent, actualChildProps]
+        [reactReduxForwardedRef, WrappedComponent, childProps]
       )
-
-      // If React sees the exact same element reference as last time, it bails out of re-rendering
-      // that child, same as if it was wrapped in React.memo() or returned false from shouldComponentUpdate.
-      const renderedChild = useMemo(() => {
-        if (shouldHandleStateChanges) {
-          // If this component is subscribed to store updates, we need to pass its own
-          // subscription instance down to our descendants. That means rendering the same
-          // Context instance, and putting a different value into the context.
-          return (
-            <ContextToUse.Provider value={overriddenContextValue}>
-              {renderedWrappedComponent}
-            </ContextToUse.Provider>
-          )
-        }
-
-        return renderedWrappedComponent
-      }, [ContextToUse, renderedWrappedComponent, overriddenContextValue])
-
-      return renderedChild
     }
 
     // If we're in "pure" mode, ensure our wrapper component only re-renders when incoming props have changed.
@@ -766,14 +524,14 @@ function connect<
         return <Connect {...props} reactReduxForwardedRef={ref} />
       })
 
-      const forwarded = _forwarded as ConnectedWrapperComponent
+      const forwarded = _forwarded as unknown as ConnectedWrapperComponent
       forwarded.displayName = displayName
       forwarded.WrappedComponent = WrappedComponent
       return hoistStatics(forwarded, WrappedComponent)
     }
 
     return hoistStatics(Connect, WrappedComponent)
-  }
+  }) as AdvancedComponentDecorator<TOwnProps, WrappedComponentProps>
 
   return wrapWithConnect
 }
