@@ -1,105 +1,10 @@
-import { useReducer, useRef, useMemo, useContext, useDebugValue } from 'react'
+import { useCallback, useContext, useDebugValue, useRef } from 'react'
 import { useReduxContext as useDefaultReduxContext } from './useReduxContext'
-import { createSubscription, Subscription } from '../utils/Subscription'
-import { useIsomorphicLayoutEffect } from '../utils/useIsomorphicLayoutEffect'
 import { ReactReduxContext } from '../components/Context'
-import { AnyAction, Store } from 'redux'
 import { DefaultRootState, EqualityFn } from '../types'
+import { useStoreSource } from '../utils/useStoreSource'
 
 const refEquality: EqualityFn<any> = (a, b) => a === b
-
-type TSelector<S, R> = (state: S) => R
-
-function useSelectorWithStoreAndSubscription<TStoreState, TSelectedState>(
-  selector: TSelector<TStoreState, TSelectedState>,
-  equalityFn: EqualityFn<TSelectedState>,
-  store: Store<TStoreState, AnyAction>,
-  contextSub: Subscription
-): TSelectedState {
-  const [, forceRender] = useReducer((s) => s + 1, 0)
-
-  const subscription = useMemo(
-    () => createSubscription(store, contextSub),
-    [store, contextSub]
-  )
-
-  const latestSubscriptionCallbackError = useRef<Error>()
-  const latestSelector = useRef<TSelector<TStoreState, TSelectedState>>()
-  const latestStoreState = useRef<TStoreState>()
-  const latestSelectedState = useRef<TSelectedState>()
-
-  const storeState = store.getState()
-  let selectedState: TSelectedState | undefined
-
-  try {
-    if (
-      selector !== latestSelector.current ||
-      storeState !== latestStoreState.current ||
-      latestSubscriptionCallbackError.current
-    ) {
-      const newSelectedState = selector(storeState)
-      // ensure latest selected state is reused so that a custom equality function can result in identical references
-      if (
-        latestSelectedState.current === undefined ||
-        !equalityFn(newSelectedState, latestSelectedState.current)
-      ) {
-        selectedState = newSelectedState
-      } else {
-        selectedState = latestSelectedState.current
-      }
-    } else {
-      selectedState = latestSelectedState.current
-    }
-  } catch (err) {
-    if (latestSubscriptionCallbackError.current) {
-      ;(
-        err as Error
-      ).message += `\nThe error may be correlated with this previous error:\n${latestSubscriptionCallbackError.current.stack}\n\n`
-    }
-
-    throw err
-  }
-
-  useIsomorphicLayoutEffect(() => {
-    latestSelector.current = selector
-    latestStoreState.current = storeState
-    latestSelectedState.current = selectedState
-    latestSubscriptionCallbackError.current = undefined
-  })
-
-  useIsomorphicLayoutEffect(() => {
-    function checkForUpdates() {
-      try {
-        const newStoreState = store.getState()
-        const newSelectedState = latestSelector.current!(newStoreState)
-
-        if (equalityFn(newSelectedState, latestSelectedState.current)) {
-          return
-        }
-
-        latestSelectedState.current = newSelectedState
-        latestStoreState.current = newStoreState
-      } catch (err) {
-        // we ignore all errors here, since when the component
-        // is re-rendered, the selectors are called again, and
-        // will throw again, if neither props nor store state
-        // changed
-        latestSubscriptionCallbackError.current = err as Error
-      }
-
-      forceRender()
-    }
-
-    subscription.onStateChange = checkForUpdates
-    subscription.trySubscribe()
-
-    checkForUpdates()
-
-    return () => subscription.tryUnsubscribe()
-  }, [store, subscription])
-
-  return selectedState!
-}
 
 /**
  * Hook factory, which creates a `useSelector` hook bound to a given context.
@@ -135,15 +40,24 @@ export function createSelectorHook(
         )
       }
     }
-    const { store, subscription: contextSub } = useReduxContext()!
+    const { storeSource } = useReduxContext()!
 
-    const selectedState = useSelectorWithStoreAndSubscription(
-      selector,
-      equalityFn,
-      store,
-      contextSub
+    const lastValue = useRef<Selected | undefined>()
+    const getSnapshot = useCallback(
+      (store): Selected => {
+        const value = selector(store.getState())
+        if (
+          lastValue.current === undefined ||
+          !equalityFn(value, lastValue.current)
+        ) {
+          lastValue.current = value
+        }
+        return lastValue.current
+      },
+      [selector, equalityFn]
     )
 
+    const selectedState = useStoreSource(storeSource, getSnapshot)
     useDebugValue(selectedState)
 
     return selectedState
